file extensions to jsx for react components. renamed dirs to functional for functional components and UI for UI components.
